Adding files for the website
Created a folder for images, added images to the folder, and created
links in the index.html file to properly display the images.
Created the lists for the proper steps as well as the ingredients
needed one ordered the other unordered.
Added all text to the recipes including desctiption, steps, and
ingredients.